+++ src/main/java/com/inq/wishhair/wesharewishhair/photo/domain/Photo.java
@@ -1,4 +1,4 @@
-package com.inq.wishhair.wesharewishhair.photo.entity;
+package com.inq.wishhair.wesharewishhair.photo.domain;
 
 import com.inq.wishhair.wesharewishhair.hairstyle.domain.HairStyle;
 import com.inq.wishhair.wesharewishhair.review.domain.Review;
