+++ src/main/java/com/inq/wishhair/wesharewishhair/photo/dto/response/PhotoResponse.java
@@ -1,6 +1,6 @@
 package com.inq.wishhair.wesharewishhair.photo.dto.response;
 
-import com.inq.wishhair.wesharewishhair.photo.entity.Photo;
+import com.inq.wishhair.wesharewishhair.photo.domain.Photo;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
