+++ src/main/java/com/inq/wishhair/wesharewishhair/review/domain/Review.java
@@ -3,7 +3,7 @@ package com.inq.wishhair.wesharewishhair.review.domain;
 import com.inq.wishhair.wesharewishhair.auditing.BaseEntity;
 import com.inq.wishhair.wesharewishhair.hairstyle.domain.HairStyle;
 import com.inq.wishhair.wesharewishhair.review.domain.likereview.LikeReview;
-import com.inq.wishhair.wesharewishhair.photo.entity.Photo;
+import com.inq.wishhair.wesharewishhair.photo.domain.Photo;
 import com.inq.wishhair.wesharewishhair.review.domain.likereview.LikeReviews;
 import com.inq.wishhair.wesharewishhair.review.enums.Score;
 import com.inq.wishhair.wesharewishhair.user.domain.User;
