+++ src/main/java/com/inq/wishhair/wesharewishhair/review/service/dto/response/ReviewResponse.java
@@ -2,7 +2,7 @@ package com.inq.wishhair.wesharewishhair.review.service.dto.response;
 
 import com.inq.wishhair.wesharewishhair.hairstyle.domain.hashtag.HashTag;
 import com.inq.wishhair.wesharewishhair.hairstyle.service.dto.response.HashTagResponse;
-import com.inq.wishhair.wesharewishhair.photo.entity.Photo;
+import com.inq.wishhair.wesharewishhair.photo.domain.Photo;
 import com.inq.wishhair.wesharewishhair.review.domain.Review;
 import com.inq.wishhair.wesharewishhair.photo.dto.response.PhotoResponse;
 import jakarta.persistence.Persistence;
